% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{.matrix_ad_pvalue}
\alias{.matrix_ad_pvalue}
\title{Approximate p-value for Anderson-Darling statistic}
\usage{
.matrix_ad_pvalue(a_sq)
}
\arguments{
\item{a_sq}{Anderson-Darling squared statistic.}
}
\description{
Approximate p-value for Anderson-Darling statistic
}
\keyword{internal}
