% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_DBSCAN.R
\name{matrix_dbscan_cluster_assignments}
\alias{matrix_dbscan_cluster_assignments}
\title{DBSCAN Cluster Assignments}
\usage{
matrix_dbscan_cluster_assignments(x_matrix, eps, minPts)
}
\arguments{
\item{x_matrix}{Numeric matrix of vectorized observations.}

\item{eps}{Neighborhood radius.}

\item{minPts}{Minimum neighborhood size.}
}
\value{
Integer vector of cluster assignments (0 = noise).
}
\description{
DBSCAN Cluster Assignments
}
\keyword{internal}
