% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_DBSCAN.R
\name{matrix_dbscan_heuristic_eps}
\alias{matrix_dbscan_heuristic_eps}
\title{DBSCAN Heuristic Epsilon}
\usage{
matrix_dbscan_heuristic_eps(x_matrix, minPts)
}
\arguments{
\item{x_matrix}{Numeric matrix of vectorized observations.}

\item{minPts}{Minimum neighborhood size.}
}
\value{
Numeric epsilon value.
}
\description{
DBSCAN Heuristic Epsilon
}
\keyword{internal}
