% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{matrix_e_step_log_density}
\alias{matrix_e_step_log_density}
\title{Compute E-Step Log-Densities for Gaussian Components}
\usage{
matrix_e_step_log_density(x_list, params, g, n)
}
\arguments{
\item{x_list}{List of matrices.}

\item{params}{Parameter list with pi, M, U, V.}

\item{g}{Number of Gaussian components.}

\item{n}{Number of observations.}
}
\value{
A matrix (n x g) of weighted log-densities for Gaussian components.
}
\description{
Evaluates log(pi_g) + log f(X_i | theta_g) for each observation and component.
}
\keyword{internal}
