% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{matrix_init_dispatch}
\alias{matrix_init_dispatch}
\title{Dispatch Initialization Scheme}
\usage{
matrix_init_dispatch(x_list, g, init, nstart = 10)
}
\arguments{
\item{x_list}{Validated list of matrices.}

\item{g}{Number of components.}

\item{init}{One of "kmeans", "emrefine", or "dbscan".}

\item{nstart}{Number of k-means restarts (used only for kmeans init).}
}
\value{
Initial parameter list (pi, M, U, V, cluster).
}
\description{
Selects and runs the appropriate initialization method.
}
\keyword{internal}
