% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_Kmeans.R
\name{matrix_initialization_loglik}
\alias{matrix_initialization_loglik}
\title{Initialization Log-Likelihood}
\usage{
matrix_initialization_loglik(params, x_list, g)
}
\arguments{
\item{params}{Initial parameter list.}

\item{x_list}{List of matrices.}

\item{g}{Number of components.}
}
\value{
Numeric log-likelihood.
}
\description{
Computes the observed-data log-likelihood for a set of initial parameters.
}
\keyword{internal}
