% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init_Kmeans.R
\name{matrix_kmeanspp_centers}
\alias{matrix_kmeanspp_centers}
\title{K-Means++ Center Seeding}
\usage{
matrix_kmeanspp_centers(x_matrix, g, n)
}
\arguments{
\item{x_matrix}{Numeric matrix of vectorized observations (n × d).}

\item{g}{Number of centers to select.}

\item{n}{Number of observations.}
}
\value{
A numeric matrix (g × d) of selected centers.
}
\description{
K-Means++ Center Seeding
}
\keyword{internal}
