% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{matrix_noise_ad_test}
\alias{matrix_noise_ad_test}
\title{Anderson-Darling Test for Matrix-Variate Normality}
\usage{
matrix_noise_ad_test(fit, x_list)
}
\arguments{
\item{fit}{A fitted model from \code{matrix_variate_noise_fit()}.}

\item{x_list}{List of matrices used for fitting.}
}
\value{
A list with component-wise Anderson-Darling test results.
}
\description{
Performs an Anderson-Darling test on Mahalanobis distances to provide
more power against alternatives in the tails compared to the KS test.
}
