% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{matrix_noise_ecdf_vs_cdf_plot}
\alias{matrix_noise_ecdf_vs_cdf_plot}
\title{Matrix Diagnostics}
\usage{
matrix_noise_ecdf_vs_cdf_plot(
  fit,
  x_list,
  main = "ECDF vs Theoretical CDF (Chi-square)",
  xlab = "Mahalanobis distance",
  ylab = "CDF",
  plot_type = c("base")
)
}
\description{
Matrix Diagnostics
}
