% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{matrix_weighted_mean}
\alias{matrix_weighted_mean}
\title{Compute Weighted Mean Matrix (M-Step)}
\usage{
matrix_weighted_mean(x_list, weights, weights_sum, r, p)
}
\arguments{
\item{x_list}{List of matrices.}

\item{weights}{Numeric vector of responsibilities for one component.}

\item{weights_sum}{Sum of weights (effective sample size).}

\item{r}{Number of rows.}

\item{p}{Number of columns.}
}
\value{
Weighted mean matrix (r x p).
}
\description{
Compute Weighted Mean Matrix (M-Step)
}
\keyword{internal}
