% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rimle_fit.R
\name{print.rimle_fit}
\alias{print.rimle_fit}
\title{Print Method for RIMLE Fits}
\usage{
\method{print}{rimle_fit}(x, ...)
}
\arguments{
\item{x}{A fitted \code{rimle_fit} object.}

\item{...}{Additional arguments (unused).}
}
\value{
\code{x} invisibly.
}
\description{
Print Method for RIMLE Fits
}
