% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rimle_fit.R
\name{print.summary.rimle_fit}
\alias{print.summary.rimle_fit}
\title{Print Summary Method for RIMLE Fits}
\usage{
\method{print}{summary.rimle_fit}(x, ...)
}
\arguments{
\item{x}{A \code{summary.rimle_fit} object.}

\item{...}{Additional arguments (unused).}
}
\value{
\code{x} invisibly.
}
\description{
Print Summary Method for RIMLE Fits
}
