% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rimle_fit.R
\name{summary.rimle_fit}
\alias{summary.rimle_fit}
\title{Summary Method for RIMLE Fits}
\usage{
\method{summary}{rimle_fit}(object, ...)
}
\arguments{
\item{object}{A fitted \code{rimle_fit} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list of class \code{"summary.rimle_fit"}.
}
\description{
Summary Method for RIMLE Fits
}
